Re-export amount from primitives
We are trying to make it so that what ever crate a user uses they see
the same module/type tree (if the module or type exists).

E.g., one can do either of these.

If they use `bitcoin`:

```
use bitcoin::{
    amount, block, fee_rate, weight, merkle_tree, opcodes,
    pow, script, sequence, transaction, witness,
};
```

Or if they use `primitives`:

```
use bitcoin_primitives::{amount, block, fee_rate, weight};
```

The above imports were tested and `amount` was found to be missing.