Fix off-by-one-bug in absolute timelock
We just fixed this in relative locktime (#4448) and forgot to check the
absolute lock time where the exact same bug exists - face palm.

When checking a locktime against block height we add 1 because when the
next block is being validated that is what the height will be and
`is_satisfied_by` is defined to return true if a transaction with this
locktime can be included in the next block.