units: Set explicit internals dependency version to v0.4.0
`cargo publish` does not allow publishing a crate without all
dependencies having explicit version numbers. We discovered recently
that having patching the root manifest was annoying when downstream
testing so we removing the version numbers in #4284 in favor of `path`.
It turns out we can include both.

Props to nyonson for discovering that.

ref: https://doc.rust-lang.org/cargo/reference/specifying-dependencies.html#multiple-locations

Use an explicit version number to the `internals` dep in `units`.