Use relative import paths rather than absolute
When doing large refactorings such as moving entire modules around
referring to items defined in a parent module by mentioning `crate` is
umbersome since the paths change and make the move confusing. It also
obscures the fact that some concepts are related.

This changes a few paths to relative so that the refactoring in the
future will be easier.