Remove BuilderExt trait
We (I) must have created the `BuilderExt` at some stage when we
thought the `Builder` would go in `primitives` - it has not and will
not for the next release so we do not need an extension trait.

Inline the `push_key` and `push_x_only_key` methods into the main
`Builder` impl block.